properly commented and cleaned P4
